requirements.txt
django
django-rest-framework
drf-spectacular[sidecar]
kolo